#1: Move everthing to new structure
